pyproject.toml
[tool.poetry]
name = "taskeasysoft"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
notebook = "^7.2.1"
jupyterlab = "^4.2.4"
fastapi = "^0.111.1"
uvicorn = "^0.30.3"
openai = "^1.37.1"
faiss-cpu = "^1.8.0.post1"
langchain = "^0.2.11"
langchain-community = "^0.2.10"
markdown = "^3.6"
black = "^24.4.2"
sentence-transformers = "^3.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
